+++ src/src/settingswidget.cpp
@@ -20,7 +20,7 @@ SettingsWidget::SettingsWidget(QWidget *parent)
 
     layout->addWidget(new QLabel("<b>Settings</b>"));
 
-    this->checkbox_retroroms = new QCheckBox("Show curated Philips P2000T ROM images");
+    this->checkbox_retroroms = new QCheckBox("Show curated Philips P2000T and P2000M ROM images");
     this->checkbox_retroroms->setChecked(this->settings.value("SHOW_RETROROMS", QVariant(true)).toBool());
     layout->addWidget(this->checkbox_retroroms);
     connect(this->checkbox_retroroms, SIGNAL(stateChanged(int)), this, SLOT(slot_settings_update(int)));